requirements.txt
mediapipe-silicon
numpy
opencv-python
opencv-contrib-python